Changes aren't reflecting
im not seeing any errors in the terminal like the tutorial.

im not seeing any css or scss updates in the browser after sync.

all code i've compared to the Coder Coder finished files are pretty matched up as it stands.

moving forward with commit hoping that just following the tutorial with present insight and rectify any issues. *shrugs